add a brief README and fix setup.py so building alone doesn't delete the build dir

+++ server/setup.py
@@ -1,10 +1,15 @@
-import os, shutil
+import os, shutil, sys
 from distutils.core import setup, Extension
 
-shutil.copyfile("convergence-notary.py", "convergence/convergence-notary")
-shutil.copyfile("convergence-gencert.py", "convergence/convergence-gencert")
-shutil.copyfile("convergence-createdb.py", "convergence/convergence-createdb")
-shutil.copyfile("convergence-bundle.py", "convergence/convergence-bundle")
+cmd = ''
+if len(sys.argv) > 1:
+    cmd = sys.argv[1]
+
+if cmd in ['build', 'install']:
+    shutil.copyfile("convergence-notary.py", "convergence/convergence-notary")
+    shutil.copyfile("convergence-gencert.py", "convergence/convergence-gencert")
+    shutil.copyfile("convergence-createdb.py", "convergence/convergence-createdb")
+    shutil.copyfile("convergence-bundle.py", "convergence/convergence-bundle")
 
 setup  (name        = 'convergence-notary',
         version     = '0.01',
@@ -20,13 +25,14 @@ setup  (name        = 'convergence-notary',
                       ('/etc/init.d', ['init-script/convergence'])]
        )
 
-try:
-    os.remove("convergence/convergence-notary")
-    os.remove("convergence/convergence-bundle")
-    os.remove('convergence/convergence-createdb')
-    os.remove("convergence/convergence-gencert")
-except:
-    pass
+if cmd in ['install', 'clean']:
+    try:
+        os.remove("convergence/convergence-notary")
+        os.remove("convergence/convergence-bundle")
+        os.remove('convergence/convergence-createdb')
+        os.remove("convergence/convergence-gencert")
+    except:
+        pass
 
-def capture(cmd):
-    return os.popen(cmd).read().strip()
+if cmd in ['install', 'clean']:
+    shutil.rmtree('build')
